/bookings/[status]
